Rename parsl.launchers.error to errors, plural, for consistency
