+++ parsl/providers/cluster_provider.py
@@ -3,7 +3,7 @@ from abc import abstractmethod
 from string import Template
 
 from parsl.providers.errors import SchedulerMissingArgs, ScriptPathError
-from parsl.launchers.error import BadLauncher
+from parsl.launchers.errors import BadLauncher
 from parsl.providers.base import ExecutionProvider
 
 logger = logging.getLogger(__name__)
